First commit for ProfilePage Feature
Implemented Friendslist JS with JSX and CSS for it.

Implemented Navbar JS with JSX and CSS for it.

Implemented ProfileData JS with JSX and CSS for it.

Implemented UserScore JS with JSX and CSS for it.

Implemented ProfilePage JS to incorporate components that go into ProfilePage and added CSS to Style